requirements.txt
django==5.0.6
djangorestframework
psycopg2-binary
drf-spectacular
dotenv